refactor: rename shared webview bridge
