+++ src/progressView/frontend/eventHandlers.ts
@@ -2,7 +2,7 @@ import { create } from 'mutative';
 
 // Local imports - shared webview
 import { PROGRESS_VIEW_COMMANDS } from '@common/webview/commands';
-import { postMessage } from '@shared/vscode';
+import { postMessage } from '@shared/hostBridge';
 import type { StreamTabId } from '@shared/schemas';
 import { PERMISSION_KIND } from '@shared/utils/uiConstants';
