+++ src/shared/commons/index.ts
@@ -71,8 +71,14 @@ export {
 // Shared: Base webview app
 export { BaseWebviewApp } from '@shared/BaseWebviewApp';
 
-// Shared: VS Code API
-export { vscode, postMessage, type VSCodeApi } from '@shared/vscode';
+// Shared: host bridge API
+export {
+  hostBridge,
+  vscode,
+  postMessage,
+  type HostBridgeApi,
+  type VSCodeApi,
+} from '@shared/hostBridge';
 
 // Shared: Schemas (re-export everything)
 export * from '@shared/schemas';
